Move test files for TestClassLoaderUtils to separate package
