+++ shared/src/test/scala/za/co/absa/hyperdrive/shared/utils/TestClassLoaderUtils.scala
@@ -20,6 +20,7 @@ package za.co.absa.hyperdrive.shared.utils
 import java.net.{URL, URLClassLoader}
 
 import org.scalatest.FlatSpec
+import za.co.absa.hyperdrive.shared.utils.classloaderutils.{TestClassLoaderUtilsNoSingleton, TestClassLoaderUtilsNotInstanceOf, TestClassLoaderUtilsSingletonInstanceOf, TestClassLoaderUtilsTestTrait}
 
 class TestClassLoaderUtils extends FlatSpec {
 
@@ -35,6 +36,7 @@ class TestClassLoaderUtils extends FlatSpec {
     assert(singleton.equals(TestClassLoaderUtilsSingletonInstanceOf))
   }
 
+  // If this test fails, you might have to check TestClassLoaderUtilsOther.jar and createTestClassLoaderUtilsOtherJar.sh
   it should "return the singleton class for a given trait and fully qualified class name from other jar" in {
     // given
     val jarPath: URL = getClass.getClassLoader.getResource("TestClassLoaderUtilsOther.jar")
