+++ shared/src/test/scala/za/co/absa/hyperdrive/shared/utils/classloaderutils/TestClassLoaderUtilsNoSingleton.scala
@@ -15,7 +15,7 @@
  *   limitations under the License.
  */
 
-package za.co.absa.hyperdrive.shared.utils
+package za.co.absa.hyperdrive.shared.utils.classloaderutils
 
 class TestClassLoaderUtilsNoSingleton {
