+++ shared/src/test/scala/za/co/absa/hyperdrive/shared/utils/classloaderutils/TestClassLoaderUtilsSingletonInstanceOf.scala
@@ -15,7 +15,7 @@
  *   limitations under the License.
  */
 
-package za.co.absa.hyperdrive.shared.utils
+package za.co.absa.hyperdrive.shared.utils.classloaderutils
 
 object TestClassLoaderUtilsSingletonInstanceOf extends TestClassLoaderUtilsTestTrait {
